improved switch targets function when locked on
Before: When looking for a new target, it would do a line trace to the
left or right multiple times.

Now: We shoot one sphere trace like a line, so that way we accomplishg
two things.
1. Reduced how many times we check to see if we hit something, which
means it's more efficient.
2. It will now hit targets above or below the player character, before
the line trace would sometimes miss targets above or below main
character.