Updated HasLineOfSight(), now checks to see if The test target is currently being rendered,
HasLineOfSight() now checks to see if The test target is currently being
rendered, if it is, go ahead and return true. If not, then go ahead and
do the trace test to see if they're behind a way of some sort.

With this, now we will be able to lock onto target that are currently
being rendered, and not having the bad luck of being at an awkward
poston where the line trace would hit something, where realsitcally we
could see them anyways.